Lag egne filer for data classes
