+++ src/main/kotlin/no/nav/melosys/skjema/controller/ArbeidsgiverRequest.kt
@@ -1,0 +1,6 @@
+package no.nav.melosys.skjema.controller
+
+data class ArbeidsgiverRequest(
+    val organisasjonsnummer: String,
+    val organisasjonNavn: String
+)
