+++ src/main/kotlin/no/nav/melosys/skjema/controller/ArbeidstakerLonnRequest.kt
@@ -1,0 +1,6 @@
+package no.nav.melosys.skjema.controller
+
+data class ArbeidstakerLonnRequest(
+    val arbeidsgiverBetalerAllLonnOgNaturaytelserIUtsendingsperioden: Boolean,
+    val virksomheterSomUtbetalerLonnOgNaturalytelser: VirksomheterSomUtbetalerLonnOgNaturalytelser?
+)
