+++ src/main/kotlin/no/nav/melosys/skjema/controller/ArbeidstakerRequest.kt
@@ -1,0 +1,16 @@
+package no.nav.melosys.skjema.controller
+
+import java.time.LocalDate
+
+data class ArbeidstakerRequest(
+    val harNorskFodselsnummer: Boolean,
+    val fodselsnummer: String?,
+    val fornavn: String?,
+    val etternavn: String?,
+    val fodselsdato: LocalDate?,
+    val harVaertEllerSkalVaereILonnetArbeidFoerUtsending: Boolean,
+    val aktivitetIMaanedenFoerUtsendingen: String,
+    val skalJobbeForFlereVirksomheter: Boolean,
+    val norskeVirksomheterArbeidstakerJobberForIutsendelsesPeriode: List<NorskVirksomhet>?,
+    val utenlandskeVirksomheterArbeidstakerJobberForIutsendelsesPeriode: List<UtenlandskVirksomhet>?
+)
