+++ src/main/kotlin/no/nav/melosys/skjema/controller/CreateSkjemaRequest.kt
@@ -1,0 +1,6 @@
+package no.nav.melosys.skjema.controller
+
+data class CreateSkjemaRequest(
+    val fnr: String,
+    val orgnr: String
+)
