+++ src/main/kotlin/no/nav/melosys/skjema/controller/NorskVirksomhet.kt
@@ -1,0 +1,5 @@
+package no.nav.melosys.skjema.controller
+
+data class NorskVirksomhet(
+    val organisasjonsnummer: String
+)
