+++ src/main/kotlin/no/nav/melosys/skjema/controller/OppsummeringRequest.kt
@@ -1,0 +1,8 @@
+package no.nav.melosys.skjema.controller
+
+import java.time.Instant
+
+data class OppsummeringRequest(
+    val bekreftetRiktighet: Boolean,
+    val submittedAt: Instant
+)
