+++ src/main/kotlin/no/nav/melosys/skjema/controller/SkatteforholdOgInntektRequest.kt
@@ -1,0 +1,9 @@
+package no.nav.melosys.skjema.controller
+
+data class SkatteforholdOgInntektRequest(
+    val erSkattepliktigTilNorgeIHeleutsendingsperioden: Boolean,
+    val mottarPengestotteFraAnnetEosLandEllerSveits: Boolean,
+    val landSomUtbetalerPengestotte: String?,
+    val pengestotteSomMottasFraAndreLandBelop: String?,
+    val pengestotteSomMottasFraAndreLandBeskrivelse: String?
+)
