+++ src/main/kotlin/no/nav/melosys/skjema/controller/VirksomhetRequest.kt
@@ -1,0 +1,7 @@
+package no.nav.melosys.skjema.controller
+
+data class VirksomhetRequest(
+    val erArbeidsgiverenOffentligVirksomhet: Boolean,
+    val erArbeidsgiverenBemanningsEllerVikarbyraa: Boolean,
+    val opprettholderArbeidsgivereVanligDrift: Boolean
+)
