+++ src/main/kotlin/no/nav/melosys/skjema/controller/VirksomheterSomUtbetalerLonnOgNaturalytelser.kt
@@ -1,0 +1,6 @@
+package no.nav.melosys.skjema.controller
+
+data class VirksomheterSomUtbetalerLonnOgNaturalytelser(
+    val norskeVirksomheter: List<NorskVirksomhet>?,
+    val utenlandskeVirksomheter: List<UtenlandskVirksomhet>?
+)
